Get types to be in types folder
