+++ client/src/component/orgamisms/DetailModal/template/index.tsx
@@ -3,31 +3,6 @@ import React from 'react';
 import CloseBtn from 'component/atoms/CloseBtn';
 import * as S from './style';
 
-export type ContentItem = {
-  title: string;
-  text: string;
-};
-export type QuestionItem = {
-  title: string;
-  answers: string[];
-};
-export type TeamListType = {
-  id: string;
-  name: string;
-};
-export type MailType = {
-  from: string;
-  teamId: string;
-  type: string;
-  teamName: string;
-  date: Date;
-};
-export type CommentsType = {
-  date: Date;
-  owner: string;
-  comment: string;
-  name: string;
-}
 interface IDetailModalTemplate {
   modalHeader: React.ReactNode;
   modalBody: React.ReactNode;
